feat(06-02): replace Analysis.tsx monolith with slim page shell
Analysis.tsx (2358 lines) → analysis/index.tsx (400 lines) + re-export shim.
Extracted data hooks into analysis/hooks.ts. Renamed context.ts → context.tsx
for JSX support. Frontend builds successfully.

+++ backend/pipeline/src/utils/bbox_utils.py
@@ -1,6 +1,6 @@
 def get_center_of_bbox(bbox):
     x1, y1, x2, y2 = bbox
-    return int((x1 + x2) / 2), int((y1 + y2) / 2)
+    return (x1 + x2) / 2.0, (y1 + y2) / 2.0
 
 
 def get_bbox_width(bbox):
@@ -9,7 +9,7 @@ def get_bbox_width(bbox):
 
 def get_foot_position(bbox):
     x1, y1, x2, y2 = bbox
-    return int((x1 + x2) / 2), int(y2)
+    return (x1 + x2) / 2.0, float(y2)
 
 
 def measure_distance(p1, p2):
